# apps/web/src/lib/service-clients/service-properties/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "properties_service",
    "description": "",
    "termsOfService": "https://macro.com/terms",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/properties/definitions": {
      "get": {
        "tags": ["Properties"],
        "summary": "List property definitions with flexible filtering",
        "operationId": "list_properties",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "description": "Filter by scope: 'user', 'team', 'system', or 'all'",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyScope"
            }
          },
          {
            "name": "include_options",
            "in": "query",
            "description": "Whether to include property options in the response",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "for_entity_type",
            "in": "query",
            "description": "Filter properties applicable to a specific entity type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EntityType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Properties retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyDefinitionResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Properties"],
        "summary": "Create a new property definition",
        "operationId": "create_property_definition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Property definition created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDefinition"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Team membership required for team scope"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/definitions/{definition_id}": {
      "get": {
        "tags": ["Properties"],
        "summary": "Get a property definition by ID",
        "operationId": "get_property_definition",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property definition retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDefinition"
                }
              }
            }
          },
          "400": {
            "description": "Invalid property ID"
          },
          "404": {
            "description": "Property definition not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": ["Properties"],
        "summary": "Delete a property definition",
        "operationId": "delete_property_definition",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Property definition deleted successfully"
          },
          "400": {
            "description": "Invalid property ID"
          },
          "403": {
            "description": "Cannot modify system properties"
          },
          "404": {
            "description": "Property definition not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/definitions/{definition_id}/options": {
      "get": {
        "tags": ["Properties"],
        "summary": "Get all options for a property (for dropdowns)",
        "operationId": "get_property_options",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property options retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyOption"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Property not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Properties"],
        "summary": "Add a new option to a property dropdown",
        "operationId": "add_property_option",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPropertyOptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Property option created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOption"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Cannot modify system properties"
          },
          "404": {
            "description": "Property not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/definitions/{definition_id}/options/{option_id}": {
      "delete": {
        "tags": ["Properties"],
        "summary": "Delete a property option",
        "operationId": "delete_property_option",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "option_id",
            "in": "path",
            "description": "Property option ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Property option deleted successfully"
          },
          "400": {
            "description": "Invalid option ID"
          },
          "403": {
            "description": "Cannot modify system properties"
          },
          "404": {
            "description": "Property option not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "tags": ["Properties"],
        "summary": "Update a property option in place (rename / recolor / reorder).",
        "description": "The option id is preserved, so the change is reflected on every entity that references it.",
        "operationId": "update_property_option",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "option_id",
            "in": "path",
            "description": "Property option ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyOptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Property option updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOption"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Cannot modify system properties"
          },
          "404": {
            "description": "Property or option not found"
          },
          "409": {
            "description": "An option with that value already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/entities/bulk": {
      "post": {
        "tags": ["Properties"],
        "summary": "Get properties for multiple entities in bulk (public endpoint with user auth)",
        "description": "Only returns properties for entities the user has view permission for.\nEntities without permission are silently omitted from the response.",
        "operationId": "get_bulk_entity_properties",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkEntityPropertiesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk entity properties retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/EntityPropertiesResponse"
                  },
                  "propertyNames": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "403": {
            "description": "Forbidden - User does not have permission"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/entities/{entity_type}/{entity_id}": {
      "get": {
        "tags": ["Properties"],
        "summary": "Get all properties for an entity",
        "operationId": "get_entity_properties",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "description": "Canonical entity type; tasks use DOCUMENT",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyTargetEntityType"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "description": "Entity ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include_metadata",
            "in": "query",
            "description": "Whether to include property metadata (default: false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entity properties retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityPropertiesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid entity type"
          },
          "403": {
            "description": "Forbidden - User does not have permission to view this entity"
          },
          "404": {
            "description": "Entity not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/entities/{entity_type}/{entity_id}/{property_id}": {
      "put": {
        "tags": ["Properties"],
        "summary": "Set or update a property value for an entity, or attach a property without a value",
        "operationId": "set_entity_property",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "description": "Canonical entity type; tasks use DOCUMENT",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyTargetEntityType"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "description": "Entity ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "property_id",
            "in": "path",
            "description": "Property ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetEntityPropertyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Entity property set successfully (with or without value)"
          },
          "400": {
            "description": "Invalid request or entity type"
          },
          "403": {
            "description": "No edit access to the entity"
          },
          "404": {
            "description": "Entity or property not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/entities/{entity_type}/{entity_id}/{property_id}/options/{option_id}": {
      "post": {
        "tags": ["Properties"],
        "summary": "Add a single option to an entity's multi-select property value.",
        "description": "Atomic delta: the option is appended to the current stored value (deduped,\nattaching the property if needed), so concurrent edits to the same value\nmerge rather than overwrite each other. Prefer this over the full-value PUT\nwhen adding one option.",
        "operationId": "add_entity_property_option",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "description": "Canonical entity type; tasks use DOCUMENT",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyTargetEntityType"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "description": "Entity ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "property_id",
            "in": "path",
            "description": "Property ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "option_id",
            "in": "path",
            "description": "Option ID to add",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Option added successfully"
          },
          "400": {
            "description": "Invalid request, property is not multi-select, or option does not belong to the property"
          },
          "403": {
            "description": "No edit access to the entity"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": ["Properties"],
        "summary": "Remove a single option from an entity's multi-select property value.",
        "description": "Atomic delta: the option is stripped from the current stored value (a no-op\nif absent), so concurrent edits to the same value merge rather than overwrite\neach other.",
        "operationId": "remove_entity_property_option",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "description": "Canonical entity type; tasks use DOCUMENT",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyTargetEntityType"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "description": "Entity ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "property_id",
            "in": "path",
            "description": "Property ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "option_id",
            "in": "path",
            "description": "Option ID to remove",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Option removed successfully"
          },
          "403": {
            "description": "No edit access to the entity"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/entity_properties/{entity_property_id}": {
      "delete": {
        "tags": ["Properties"],
        "summary": "Remove a specific entity property by its ID",
        "operationId": "delete_entity_property",
        "parameters": [
          {
            "name": "entity_property_id",
            "in": "path",
            "description": "Entity Property ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Entity property removed successfully"
          },
          "403": {
            "description": "Property is required and cannot be removed, or no edit access"
          },
          "404": {
            "description": "Entity property not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/tags": {
      "get": {
        "tags": ["Properties"],
        "summary": "List the caller's tag sets: their personal set, plus their team's set when on a team.\nPure read - a scope with no provisioned definition yet returns an empty set.",
        "operationId": "list_tags",
        "responses": {
          "200": {
            "description": "Tag sets retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagSetResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Properties"],
        "summary": "Provision (get-or-create) the caller's tag set for a scope and return it. Called when\nthe caller creates their first label for that scope, so the read path stays side-effect free.",
        "operationId": "ensure_tag_set",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnsureTagSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tag set provisioned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagSetResponse"
                }
              }
            }
          },
          "403": {
            "description": "Team membership required for team scope"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AddPropertyOptionRequest": {
        "oneOf": [
          {
            "type": "object",
            "description": "Add a string option to a SelectString property",
            "required": ["option", "type"],
            "properties": {
              "option": {
                "$ref": "#/components/schemas/AddStringOptionRequest"
              },
              "type": {
                "type": "string",
                "enum": ["select_string"]
              }
            }
          },
          {
            "type": "object",
            "description": "Add a number option to a SelectNumber property",
            "required": ["option", "type"],
            "properties": {
              "option": {
                "$ref": "#/components/schemas/AddNumberOptionRequest"
              },
              "type": {
                "type": "string",
                "enum": ["select_number"]
              }
            }
          }
        ],
        "description": "Enum for type-safe property option addition requests."
      },
...
      "BulkEntityPropertiesRequest": {
        "type": "object",
        "description": "Request for getting properties for multiple entities in bulk",
        "required": ["entities"],
        "properties": {
          "entities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyTargetReference"
            },
            "description": "Array of entity references (entity_id and entity_type pairs)"
          },
          "property_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Optional: only return properties with these definition IDs. If empty, returns all."
          }
        }
      },
      "CreatePropertyDefinitionRequest": {
        "type": "object",
        "description": "Request to create a new property definition.",
        "required": ["scope", "display_name", "data_type"],
        "properties": {
          "data_type": {
            "$ref": "#/components/schemas/PropertyDataType"
          },
          "display_name": {
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/CreatePropertyScope"
          }
        }
      },
...
      "EntityPropertiesResponse": {
        "type": "object",
        "description": "Response for document/entity properties endpoint.",
        "required": ["entity_id", "properties"],
        "properties": {
          "entity_id": {
            "type": "string"
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityPropertyWithDefinition"
            }
          }
        }
      },
...
      "EntityType": {
        "type": "string",
        "description": "Type of entity that can be referenced by entity properties.",
        "enum": [
          "CALL_RECORD",
          "CHANNEL",
          "CHAT",
          "COMPANY",
          "DOCUMENT",
          "PROJECT",
          "TASK",
          "THREAD",
          "USER"
        ]
      },
...
      "PropertyDefinition": {
        "type": "object",
        "description": "Property definition model (service representation).",
        "required": [
          "id",
          "owner",
          "display_name",
          "data_type",
          "is_multi_select",
          "created_at",
          "updated_at",
          "is_system",
          "is_metadata"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "data_type": {
            "$ref": "#/components/schemas/DataType"
          },
          "display_name": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_metadata": {
            "type": "boolean",
            "description": "Flag to indicate if this is a system-generated metadata property.\nNot stored in database - computed at service layer."
          },
          "is_multi_select": {
            "type": "boolean"
          },
          "is_system": {
            "type": "boolean",
            "description": "Flag to indicate if this is a system property (stored in DB)."
          },
          "owner": {
            "$ref": "#/components/schemas/PropertyOwner"
          },
          "specific_entity_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EntityType"
              }
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PropertyDefinitionResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PropertyDefinition"
          },
          {
            "$ref": "#/components/schemas/PropertyDefinitionWithOptions"
          }
        ],
        "description": "Response for property definition with optional property options"
      },
...
      "PropertyOption": {
        "type": "object",
        "description": "A selectable option for select-type properties (service representation).",
        "required": [
          "id",
          "property_definition_id",
          "display_order",
          "value",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "color": {
            "type": ["string", "null"]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_order": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "property_definition_id": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "$ref": "#/components/schemas/PropertyOptionValue"
          }
        }
      },
...
      "PropertyScope": {
        "type": "string",
        "description": "Scope filter for property queries",
        "enum": ["user", "team", "system", "all"]
      },
...
      "UpdatePropertyOptionRequest": {
        "type": "object",
        "description": "Request to update a property option in place. Omitted fields keep their current value.\nThe option id is preserved, so the change propagates to every entity referencing it.",
        "properties": {
          "color": {
            "type": ["string", "null"],
            "description": "New color as a hex string like `#RRGGBB`."
          },
          "display_order": {
            "type": ["integer", "null"],
            "format": "int32",
            "description": "New display order."
          },
          "value": {
            "type": ["string", "null"],
            "description": "New string value (label text). Only valid for SelectString and Tag properties."
          }
        }
      }
    }
  },
...
}